Frontend requests can be proxied to backend
Backend is now on a different port and api calls
now get proxied to that port. This allows the react
app and express server to be run separately in two
different terminals with "npm run start-client" and
"npm run start-server" respectively. This means no
more fiddling in the command line every time you
want to see a change while testing.